backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
celery==5.4.0
redis==5.2.0
pytest==8.3.3
httpx==0.27.2
PyYAML==6.0.2
